get ethereum address for burner, chain validation, reorg examples, and working ethereum example for burner
