+++ packages/bitcore-hardware/examples/ledger/address.ts
@@ -2,7 +2,11 @@ import 'source-map-support/register.js';
 import Ledger from '../../src/ledger/wallet.js';
 
 const args = process.argv.slice(2);
-const chain = args[0] || 'BTC';
+const chain = args[0]?.toUpperCase() || 'BTC';
+
+if (!Ledger.isValidChain(chain)) {
+  throw new Error(`Invalid chain: ${chain}`);
+}
 
 const ledger = new Ledger();
 await ledger.connect();
