+++ packages/bitcore-hardware/package.json
@@ -15,14 +15,11 @@
     "test": "npm run compile && mocha build/test",
     "burner:public-key": "npm run compile && node build/examples/burner/publicKey.js",
     "burner:address": "npm run compile && node build/examples/burner/address.js",
-    "burner:sign-btc": "npm run compile && node build/examples/burner/sign-btc.js",
-    "burner:sign-eth": "npm run compile && node build/examples/burner/sign-eth.js",
+    "burner:sign": "npm run compile && node build/examples/burner/sign.js",
     "burner:version": "npm run compile && node build/examples/burner/version.js",
     "ledger:public-key": "npm run compile && node build/examples/ledger/publicKey.js",
     "ledger:address": "npm run compile && node build/examples/ledger/address.js",
-    "ledger:sign-btc": "npm run compile && node build/examples/ledger/sign-btc.js",
-    "ledger:sign-eth": "npm run compile && node build/examples/ledger/sign-eth.js",
-    "ledger:sign-sol": "npm run compile && node build/examples/ledger/sign-sol.js",
+    "ledger:sign": "npm run compile && node build/examples/ledger/sign.js",
     "ledger:version": "npm run compile && node build/examples/ledger/version.js"
   },
   "keywords": [
