+++ packages/bitcore-hardware/src/types/burnerTypes.ts
@@ -4,3 +4,9 @@ export type DataType =
   'keySlotFlag' | 'keySlotFlag' | 'keySlotFailedAuthCtr' |
   'keySlotFailState' | 'keySlotAuthUnlockChallenge' | 'latchValue' |
   'latchAttest' | 'graffiti' | 'firmwareVersion';
+
+export type CommandNameType = 
+  'sign' | 'sign_random' | 'sign_challenge' |
+  'write_latch' | 'cfg_ndef' | 'gen_key' |
+  'gen_key_confirm' | 'get_pkeys' | 'get_key_info' |
+  'set_password' | 'unset_password' | 'get_data_struct_v2'
