--- packages/bitcore-hardware/src/types/paramTypes.ts
@@ -1,3 +1,0 @@
-export interface BaseParams { index: number };
-
-export type SignParams = BaseParams & { tx: any; password?: string };
